Second update
Second update on the code. This time, i started the user registration part of the code. Im kinda new in C++, so i down really now if using this much fuctions is a bad thing or not.
Tomorow im going to show my coworkers this code and gatter some ideas to make it better